requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-Session==0.5.0

# AI integration
together>=1.0.0

# Environment management
python-dotenv==1.0.0

# PDF generation for export (optional)
reportlab==4.0.4

# Data processing
pandas==2.1.1

# Development dependencies (optional)
pytest==7.4.2
pytest-flask==1.2.0

# Note: uuid, datetime, logging, json are built-in Python modules
# They don't need to be installed separately